pyproject.toml
[project]
name = "real-statistic"
version = "0.1.0"
description = "Real purchasing power comparison across RO, FR, DE; basket-based experiment challenging the PPS narrative"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
real-statistic = "real_statistic.cli:main"
